Update README.md
We take advantage of model serializers to simplify code and avoid repeating information that is already included in the model. Create a new version of the existing ToySerializer class that inherits from the rest_framework.serializers.ModelSerializer superclass.
The ModelSerializer class automatically populates a set of default fields and default validators by retrieving metadata from the related model class that we must specify. The ModelSerializer class also provides default implementations for the create and update methods. The types and max_length values related to the toy fields are only included in the Toy class.
Previously the view functions return a JSONResponse when it is necessary to return JSON data and a django.Http.Response.HttpResponse instance when the response is just an HTTP status code. Take advantage of additional features included in the Django REST framework that will make it easy for us to support additional formats for our RESTful Web Service (@api_view decorator that is declared in the rest_framework.decorators module). We will apply this decorator to our function-based views: toys_list and toys_detail. This decorator is a wrapper that converts a function-based view into a subclass of the rest_framework.views.APIView class. This class is the base class for all the views in the Django REST framework.  The usage of the decorator makes sure that we always receive an instance of the rest_framework.request.Request class in the request argument when Django calls our view function. The new code removes the JSONResponse class and uses the more generic rest_framework.response.Response class.
By default, the value for the DEFAULT_CONTENT_NEGOTIATION_CLASS is the rest_framework.negotiation.DefaultContentNegotiation class. When we use the decorator, the web service will use this content negotiation class to select the appropriate renderer for the response, based on the incoming request.
In the old version of the code, we used the JSONResponse and HttpResponse classes in the functions. The new version replaced the usages of both classes with the rest_framework.response.Response class. This way, the code takes advantage of the content negotiation features.